+++ api/Dockerfile
@@ -1,11 +1,8 @@
 FROM node:latest
 
+COPY . /web
 WORKDIR /web
 
-COPY package.json package-lock.json ./
 RUN npm install
-
-COPY . .
-RUN npm run-script build
 
 CMD npm start
