+++ api/package.json
@@ -2,9 +2,10 @@
 	"name": "gql_e1",
 	"version": "1.0.0",
 	"private": true,
+	"main": "index.js",
 	"scripts": {
-		"build": "tsc",
-		"start": "node bin/index.js"
+		"prestart": "tsc",
+		"start": "node index.js"
 	},
 	"dependencies": {
 		"@types/graphql": "^14.5.0",
