+++ api/session-resolver.ts
@@ -1,6 +1,6 @@
-import { Arg, Int, Mutation, Query, Resolver } from "type-graphql"
-import { getRepository, Repository } from "typeorm"
-import { Session } from "../entities/session"
+import { Query, Resolver, Mutation, Args, Arg, Int } from "type-graphql"
+import { Repository, getRepository} from "typeorm"
+import { Session } from "./session"
 
 @Resolver(() => Session)
 export class SessionResolver {
@@ -9,17 +9,17 @@ export class SessionResolver {
 	@Query(() => [Session])
 	protected async sessions(): Promise<Session[]> {
 		return this.sessionRepo.find()
-	}
-
+    }
+    
 	@Mutation(() => Session)
 	protected async createSession(
-		@Arg("purpose", () => String) purpose: string,
-		@Arg("description", () => String) description: string
+        @Arg("purpose", () => String) purpose: string,
+        @Arg("description", () => String) description: string 
 	): Promise<Session> {
 		const session = this.sessionRepo.create({
-			purpose: purpose,
-			description: description
-		})
+            purpose: purpose,
+            description: description
+		});
 		return session.save()
 	}
 
@@ -29,7 +29,7 @@ export class SessionResolver {
 	): Promise<boolean> {
 		await this.sessionRepo.delete({
 			id: id
-		})
+		});
 		return true
 	}
 }
